Reorganize the line items & tests
